backend/pyproject.toml
[project]
name = "cronosmatic-backend"
version = "0.1.0"
description = "CronosMatic API – FastAPI backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic-settings>=2.7",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "pgvector>=0.3",
    "alembic>=1.14",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "email-validator>=2.3.0",
    "boto3>=1.42.62",
    "python-multipart>=0.0.22",
]

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.25",
    "httpx>=0.28",
    "ruff>=0.9",
    "pytest-cov>=7.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
